refactor(chord): 获取弦图的默认配置项
